fix: resolve Kotlin compilation errors and Rust clippy warnings (76 errors)

+++ android/app/src/main/java/com/taskmod/app/MagiskGuideActivity.kt
@@ -314,7 +314,7 @@ class MagiskGuideActivity : AppCompatActivity() {
                             val progress = (bytesDownloaded * 100 / bytesTotal)
                             downloadProgress.isIndeterminate = false
                             downloadProgress.progress = progress
-                            downloadProgressText.text = formatFileSize(bytesDownloaded) + " / " + formatFileSize(bytesTotal)
+                            downloadProgressText.text = formatFileSize(bytesDownloaded.toLong()) + " / " + formatFileSize(bytesTotal.toLong())
                         }
                         cursor.close()
                         handler.postDelayed(this, 500)
